backend/requirements.txt
APScheduler==3.11.1
Django>=5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
psutil==7.1.3
tzlocal==5.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
